# generated-openapi-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Trieve API",
    "description": "Trieve OpenAPI Specification. This document describes all of the operations available through the Trieve API.",
    "contact": {
      "name": "Trieve Team",
      "url": "https://trieve.ai",
      "email": "developers@trieve.ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/devflowinc/actix-web-template/blob/main/LICENSE.txt"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/emails": {
      "post": {
        "tags": [
          "emails"
        ],
        "operationId": "create_email",
        "requestBody": {
          "description": "JSON request payload to create a new email",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmailReqPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "JSON body representing the email that was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Org"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      }
    },
    "/api/emails/{email_id}": {
      "get": {
        "tags": [
          "emails"
        ],
        "operationId": "get_email",
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "description": "The email id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON object representing the requested email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Org"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "emails"
        ],
        "operationId": "update_email",
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "description": "The email id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "JSON request payload to update the email",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmailReqPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Object representing the renamed email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Org"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "emails"
        ],
        "operationId": "delete_email",
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "description": "The email id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content response indicating that the email was successfully deleted"
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      }
    },
...
    "/api/links": {
      "post": {
        "tags": [
          "links"
        ],
        "operationId": "create_link",
        "requestBody": {
          "description": "JSON request payload to create a new link",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkReqPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "JSON body representing the link that was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Org"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      }
    },
    "/api/links/{link_id}": {
      "get": {
        "tags": [
          "links"
        ],
        "operationId": "get_link",
        "parameters": [
          {
            "name": "link_id",
            "in": "path",
            "description": "The link id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON object representing the requested link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Org"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "links"
        ],
        "operationId": "update_link",
        "parameters": [
          {
            "name": "link_id",
            "in": "path",
            "description": "The link id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "JSON request payload to update the link",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLinkReqPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Object representing the renamed link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Org"
                }
              }
            }
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "links"
        ],
        "operationId": "delete_link",
        "parameters": [
          {
            "name": "link_id",
            "in": "path",
            "description": "The link id to use for the request",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content response indicating that the link was successfully deleted"
          },
          "401": {
            "description": "Service error relating to authentication status of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespPayload"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [
              "readonly"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEmailReqPayload": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "CreateLinkReqPayload": {
        "type": "object",
        "required": [
          "link"
        ],
        "properties": {
          "link": {
            "type": "string"
          }
        }
      },
...
      "ErrorRespPayload": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "example": {
          "message": "Bad Request"
        }
      },
...
      "Org": {
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "$ref": "#/components/schemas/PrefixedUuid"
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "created_at": "2021-01-01T00:00:00",
          "id": "e3e3e3e3-e3e3-e3e3-e3e3-e3e3e3e3e3e3",
          "name": "Trieve Team",
          "updated_at": "2021-01-01T00:00:00"
        }
      },
      "PrefixedUuid": {
        "type": "string"
      },
...
      "UpdateEmailReqPayload": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateLinkReqPayload": {
        "type": "object",
        "properties": {
          "link": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  },
...
}